Move display welcome screen function to Game class
