+++ art.py
@@ -166,3 +166,29 @@ snake = [
             ██████████▒▒██░░████▒▒████████
                 ████████████████████
 {Fore.RESET}''']
+
+
+welcome = f'''
+
+
+
+                         Welcome to
+
+{Fore.LIGHTGREEN_EX}
+    ███╗   ███╗██╗   ██╗    ██╗     ██╗████████╗████████╗██╗     ███████╗
+    ████╗ ████║╚██╗ ██╔╝    ██║     ██║╚══██╔══╝╚══██╔══╝██║     ██╔════╝
+    ██╔████╔██║ ╚████╔╝     ██║     ██║   ██║      ██║   ██║     █████╗
+    ██║╚██╔╝██║  ╚██╔╝      ██║     ██║   ██║      ██║   ██║     ██╔══╝
+    ██║ ╚═╝ ██║   ██║       ███████╗██║   ██║      ██║   ███████╗███████╗
+    ╚═╝     ╚═╝   ╚═╝       ╚══════╝╚═╝   ╚═╝      ╚═╝   ╚══════╝╚══════╝
+
+{Style.BRIGHT}
+                    ███████╗███╗   ██╗███████╗██╗  ██╗
+                    ██╔════╝████╗  ██║██╔════╝██║ ██╔╝
+                    ███████╗██╔██╗ ██║█████╗  █████╔╝
+                    ╚════██║██║╚██╗██║██╔══╝  ██╔═██╗
+                    ███████║██║ ╚████║███████╗██║  ██╗
+                    ╚══════╝╚═╝  ╚═══╝╚══════╝╚═╝  ╚═╝
+{Style.RESET_ALL}
+
+'''
