+++ run.py
@@ -3,42 +3,42 @@ from game import Game
 from colorama import Fore, Back, Style
 
 
-def display_welcome_screen():
-    """Displays the welcome screen to the player"""
-    Game.clear()
-    print(f'''
+# def display_welcome_screen():
+#     """Displays the welcome screen to the player"""
+#     Game.clear()
+#     print(f'''
 
 
 
-                         Welcome to
+#                          Welcome to
 
-{Fore.LIGHTGREEN_EX}
-    ███╗   ███╗██╗   ██╗    ██╗     ██╗████████╗████████╗██╗     ███████╗
-    ████╗ ████║╚██╗ ██╔╝    ██║     ██║╚══██╔══╝╚══██╔══╝██║     ██╔════╝
-    ██╔████╔██║ ╚████╔╝     ██║     ██║   ██║      ██║   ██║     █████╗
-    ██║╚██╔╝██║  ╚██╔╝      ██║     ██║   ██║      ██║   ██║     ██╔══╝
-    ██║ ╚═╝ ██║   ██║       ███████╗██║   ██║      ██║   ███████╗███████╗
-    ╚═╝     ╚═╝   ╚═╝       ╚══════╝╚═╝   ╚═╝      ╚═╝   ╚══════╝╚══════╝
+# {Fore.LIGHTGREEN_EX}
+#     ███╗   ███╗██╗   ██╗    ██╗     ██╗████████╗████████╗██╗     ███████╗
+#     ████╗ ████║╚██╗ ██╔╝    ██║     ██║╚══██╔══╝╚══██╔══╝██║     ██╔════╝
+#     ██╔████╔██║ ╚████╔╝     ██║     ██║   ██║      ██║   ██║     █████╗
+#     ██║╚██╔╝██║  ╚██╔╝      ██║     ██║   ██║      ██║   ██║     ██╔══╝
+#     ██║ ╚═╝ ██║   ██║       ███████╗██║   ██║      ██║   ███████╗███████╗
+#     ╚═╝     ╚═╝   ╚═╝       ╚══════╝╚═╝   ╚═╝      ╚═╝   ╚══════╝╚══════╝
 
-{Style.BRIGHT}
-                    ███████╗███╗   ██╗███████╗██╗  ██╗
-                    ██╔════╝████╗  ██║██╔════╝██║ ██╔╝
-                    ███████╗██╔██╗ ██║█████╗  █████╔╝
-                    ╚════██║██║╚██╗██║██╔══╝  ██╔═██╗
-                    ███████║██║ ╚████║███████╗██║  ██╗
-                    ╚══════╝╚═╝  ╚═══╝╚══════╝╚═╝  ╚═╝
-{Style.RESET_ALL}
+# {Style.BRIGHT}
+#                     ███████╗███╗   ██╗███████╗██╗  ██╗
+#                     ██╔════╝████╗  ██║██╔════╝██║ ██╔╝
+#                     ███████╗██╔██╗ ██║█████╗  █████╔╝
+#                     ╚════██║██║╚██╗██║██╔══╝  ██╔═██╗
+#                     ███████║██║ ╚████║███████╗██║  ██╗
+#                     ╚══════╝╚═╝  ╚═══╝╚══════╝╚═╝  ╚═╝
+# {Style.RESET_ALL}
 
-''')
-    choice = input(f'''
-    Do you want to start a
-    {Fore.LIGHTGREEN_EX}N{Fore.RESET}EW GAME or
-    {Fore.LIGHTGREEN_EX}C{Fore.RESET}ONTINUE an existing game
+# ''')
+#     choice = input(f'''
+#     Do you want to start a
+#     {Fore.LIGHTGREEN_EX}N{Fore.RESET}EW GAME or
+#     {Fore.LIGHTGREEN_EX}C{Fore.RESET}ONTINUE an existing game
 
-    or do you want to
-    {Fore.LIGHTGREEN_EX}R{Fore.RESET}EAD the tutorial?\n\n
-    ''').lower()
-    return choice
+#     or do you want to
+#     {Fore.LIGHTGREEN_EX}R{Fore.RESET}EAD the tutorial?\n\n
+#     ''').lower()
+#     return choice
 
 
 def display_tutorial():
@@ -74,7 +74,7 @@ def display_tutorial():
 
 def main():
     """Main function for the game"""
-    game_choice = display_welcome_screen()
+    game_choice = Game.display_welcome_screen()
     if game_choice == 'n':
         my_game = Game()
         my_game.start_new_game()
